Separating asset packaging logic into a separate class that is reinstantiated on each request
